Solving warnings for Wazuh server and windows agent

+++ src/monitord/rotate_log.c
@@ -208,8 +208,11 @@ void remove_old_logs(const char *base_dir, int keep_log_days) {
         }
 
         if (sscanf(dirent->d_name, "%d", &year) > 0) {
-            snprintf(path, PATH_MAX, "%s/%s", base_dir, dirent->d_name);
-            remove_old_logs_y(path, year, threshold);
+            int size = snprintf(path, PATH_MAX, "%s/%s", base_dir, dirent->d_name);
+
+            if (size + 1 <= PATH_MAX) {
+                remove_old_logs_y(path, year, threshold);
+            }
         }
     }
 
@@ -241,12 +244,14 @@ void remove_old_logs_y(const char * base_dir, int year, time_t threshold) {
             }
         }
 
-        snprintf(path, PATH_MAX, "%s/%s", base_dir, dirent->d_name);
+        int size = snprintf(path, PATH_MAX, "%s/%s", base_dir, dirent->d_name);
 
-        if (month < 12) {
-            remove_old_logs_m(path, year, month, threshold);
-        } else {
-            mwarn("Unexpected folder '%s'", path);
+        if (size + 1 <= PATH_MAX) {
+            if (month < 12) {
+                remove_old_logs_m(path, year, month, threshold);
+            } else {
+                mwarn("Unexpected folder '%s'", path);
+            }
         }
     }
 
@@ -261,7 +266,7 @@ void remove_old_logs_m(const char * base_dir, int year, int month, time_t thresh
     time_t now = time(NULL);
     struct tm tm = { .tm_sec = 0 };
     int counter;
-
+    int size = 0;
     localtime_r(&now, &tm);
 
     tm.tm_year = year - 1900;
@@ -285,9 +290,12 @@ void remove_old_logs_m(const char * base_dir, int year, int month, time_t thresh
             tm.tm_mday = day;
 
             if (mktime(&tm) <= threshold) {
-                snprintf(path, PATH_MAX, "%s/%s", base_dir, dirent->d_name);
-                mdebug2("Removing old log '%s'", path);
-                unlink(path);
+                size = snprintf(path, PATH_MAX, "%s/%s", base_dir, dirent->d_name);
+
+                if (size + 1 <= PATH_MAX) {
+                    mdebug2("Removing old log '%s'", path);
+                    unlink(path);
+                }
             }
         }
 
@@ -295,9 +303,12 @@ void remove_old_logs_m(const char * base_dir, int year, int month, time_t thresh
             tm.tm_mday = day;
 
             if (mktime(&tm) <= threshold) {
-                snprintf(path, PATH_MAX, "%s/%s", base_dir, dirent->d_name);
-                mdebug2("Removing old log '%s'", path);
-                unlink(path);
+                size = snprintf(path, PATH_MAX, "%s/%s", base_dir, dirent->d_name);
+
+                if (size + 1 <= PATH_MAX) {
+                    mdebug2("Removing old log '%s'", path);
+                    unlink(path);
+                }
             }
         }
 
@@ -305,9 +316,12 @@ void remove_old_logs_m(const char * base_dir, int year, int month, time_t thresh
             tm.tm_mday = day;
 
             if (mktime(&tm) <= threshold) {
-                snprintf(path, PATH_MAX, "%s/%s", base_dir, dirent->d_name);
-                mdebug2("Removing old log '%s'", path);
-                unlink(path);
+                size = snprintf(path, PATH_MAX, "%s/%s", base_dir, dirent->d_name);
+
+                if (size + 1 <= PATH_MAX) {
+                    mdebug2("Removing old log '%s'", path);
+                    unlink(path);
+                }
             }
         }
 
@@ -315,9 +329,12 @@ void remove_old_logs_m(const char * base_dir, int year, int month, time_t thresh
             tm.tm_mday = day;
 
             if (mktime(&tm) <= threshold) {
-                snprintf(path, PATH_MAX, "%s/%s", base_dir, dirent->d_name);
-                mdebug2("Removing old log '%s'", path);
-                unlink(path);
+                size = snprintf(path, PATH_MAX, "%s/%s", base_dir, dirent->d_name);
+
+                if (size + 1 <= PATH_MAX) {
+                    mdebug2("Removing old log '%s'", path);
+                    unlink(path);
+                }
             }
         }
